Move frameworks to separate folder as file is too big
